rework testing a bit
