+++ features/support/hooks.ts
@@ -1,8 +1,6 @@
 import { After, Before, formatterHelpers, ITestCaseHookParameter } from '../../'
-import fs from 'fs'
 import fsExtra from 'fs-extra'
 import path from 'path'
-import tmp from 'tmp'
 import { doesHaveValue } from '../../src/value_checker'
 import { World } from './world'
 import { warnUserAboutEnablingDeveloperMode } from './warn_user_about_enabling_developer_mode'
@@ -52,57 +50,6 @@ Before('@esm', function (this: World) {
     name: 'feature-test-pickle',
     type: 'module',
   })
-})
-
-Before('@global-install', function (this: World) {
-  const tmpObject = tmp.dirSync({ unsafeCleanup: true })
-
-  // Symlink everything in node_modules so the fake global install has all the dependencies it needs
-  const projectNodeModulesPath = path.join(projectPath, 'node_modules')
-  const projectNodeModulesDirs = fs.readdirSync(projectNodeModulesPath)
-  const globalInstallNodeModulesPath = path.join(tmpObject.name, 'node_modules')
-  projectNodeModulesDirs.forEach((nodeModuleDir) => {
-    let pathsToLink = [nodeModuleDir]
-    if (nodeModuleDir[0] === '@') {
-      const scopeNodeModuleDirs = fs.readdirSync(
-        path.join(projectNodeModulesPath, nodeModuleDir)
-      )
-      pathsToLink = scopeNodeModuleDirs.map((x) => path.join(nodeModuleDir, x))
-    }
-    pathsToLink.forEach((pathToLink) => {
-      const globalInstallNodeModulePath = path.join(
-        globalInstallNodeModulesPath,
-        pathToLink
-      )
-      const projectNodeModulePath = path.join(
-        projectNodeModulesPath,
-        pathToLink
-      )
-      fsExtra.ensureSymlinkSync(
-        projectNodeModulePath,
-        globalInstallNodeModulePath
-      )
-    })
-  })
-
-  const globalInstallCucumberPath = path.join(
-    globalInstallNodeModulesPath,
-    '@cucumber',
-    'cucumber'
-  )
-  const itemsToCopy = ['bin', 'lib', 'package.json']
-  itemsToCopy.forEach((item) => {
-    fsExtra.copySync(
-      path.join(projectPath, item),
-      path.join(globalInstallCucumberPath, item)
-    )
-  })
-
-  this.globalExecutablePath = path.join(
-    globalInstallCucumberPath,
-    'bin',
-    'cucumber.js'
-  )
 })
 
 After(async function (this: World) {
